made the page into comp and also added the new magic profile
